Adding missed changes from shared citizen-sidebar component start implementation
