+++ web/src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
 
-import { AppSidebar } from '@/components/organisms/app-sidebar'
+import { AppSidebar } from '@/components/organisms/citizen-sidebar'
 
 const inter = Inter({
   variable: '--font-inter',
